requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
backports.asyncio.runner==1.2.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
blockbuster==1.5.25
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
databricks-sdk==0.59.0
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.0
forbiddenfruit==0.1.4
freezegun==1.5.3
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gto==1.7.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kombu==5.5.4
-e git+https://github.com/mdgolammafuz/langchain.git@7a8a51edccb59346b0ceae9146a6dabe09d11b06#egg=langchain&subdirectory=libs/langchain
-e git+https://github.com/mdgolammafuz/langchain.git@7a8a51edccb59346b0ceae9146a6dabe09d11b06#egg=langchain_core&subdirectory=libs/core
-e git+https://github.com/mdgolammafuz/langchain.git@7a8a51edccb59346b0ceae9146a6dabe09d11b06#egg=langchain_monorepo
langchain-text-splitters==0.3.8
langsmith==0.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
omegaconf==2.3.0
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-benchmark==5.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.4
scikit-learn==1.7.1
scipy==1.15.3
scmrepo==3.4.0
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.47.2
sympy==1.14.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.31.2
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
zstandard==0.23.0
kafka-python
